+++ core/server/api_container/server/startosis_engine/startosis_modules/git_module_content_provider/git_module_content_provider.go
@@ -9,19 +9,20 @@ import (
 )
 
 const (
-	moduleDirPermission     = 0755
-	temporaryRepoDirPattern = "tmp-repo-dir-*"
+	moduleDirPermission           = 0755
+	temporaryRepoDirPattern       = "tmp-repo-dir-*"
+	authorAndModuleNameAdjustment = 2
 )
 
 type GitModuleContentProvider struct {
-	moduleTmpDir string
-	moduleDir    string
+	modulesTmpDir string
+	modulesDir    string
 }
 
 func NewGitModuleContentProvider(moduleDir string, tmpDir string) *GitModuleContentProvider {
 	return &GitModuleContentProvider{
-		moduleDir:    moduleDir,
-		moduleTmpDir: tmpDir,
+		modulesDir:    moduleDir,
+		modulesTmpDir: tmpDir,
 	}
 }
 
@@ -31,10 +32,10 @@ func (provider *GitModuleContentProvider) GetModuleContents(moduleURL string) (s
 		return "", stacktrace.Propagate(err, "An error occurred while parsing URL '%v'", moduleURL)
 	}
 
-	pathToStartosisFile := path.Join(provider.moduleDir, parsedURL.relativeFilePath)
+	pathToFile := path.Join(provider.modulesDir, parsedURL.relativeFilePath)
 
 	// Load the file if it already exists
-	contents, err := os.ReadFile(pathToStartosisFile)
+	contents, err := os.ReadFile(pathToFile)
 	if err == nil {
 		return string(contents), nil
 	}
@@ -46,9 +47,9 @@ func (provider *GitModuleContentProvider) GetModuleContents(moduleURL string) (s
 	}
 
 	// Load it after cloning
-	contents, err = os.ReadFile(pathToStartosisFile)
+	contents, err = os.ReadFile(pathToFile)
 	if err != nil {
-		return "", stacktrace.Propagate(err, "An error occurred in reading contents of the Startosis file '%v'", pathToStartosisFile)
+		return "", stacktrace.Propagate(err, "An error occurred in reading contents of the file '%v'", pathToFile)
 	}
 
 	return string(contents), nil
@@ -58,7 +59,7 @@ func (provider *GitModuleContentProvider) GetModuleContents(moduleURL string) (s
 // TODO make this support versioning via tags, commit hashes or branches
 func (provider *GitModuleContentProvider) atomicClone(parsedURL *ParsedGitURL) error {
 	// First we clone into a temporary directory
-	tempRepoDirPath, err := os.MkdirTemp(provider.moduleTmpDir, temporaryRepoDirPattern)
+	tempRepoDirPath, err := os.MkdirTemp(provider.modulesTmpDir, temporaryRepoDirPattern)
 	if err != nil {
 		return stacktrace.Propagate(err, "Error creating temporary directory for the repository to be cloned into")
 	}
@@ -70,8 +71,8 @@ func (provider *GitModuleContentProvider) atomicClone(parsedURL *ParsedGitURL) e
 	}
 
 	// Then we move it into the target directory
-	moduleAuthorPath := path.Join(provider.moduleDir, parsedURL.moduleAuthor)
-	modulePath := path.Join(provider.moduleDir, parsedURL.relativeRepoPath)
+	moduleAuthorPath := path.Join(provider.modulesDir, parsedURL.moduleAuthor)
+	modulePath := path.Join(provider.modulesDir, parsedURL.relativeRepoPath)
 	fileMode, err := os.Stat(moduleAuthorPath)
 	if err == nil && !fileMode.IsDir() {
 		return stacktrace.Propagate(err, "Expected '%v' to be a directory but it is something else", moduleAuthorPath)